#586: Add permission constants and adopt server console
git-svn-id: https://src.heuristiclab.com/svn/core/trunk@1648 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ sources/HeuristicLab.Hive.Server.Core/3.2/ServerConsoleFacade.cs
@@ -26,6 +26,7 @@ using System.Text;
 using HeuristicLab.Hive.Contracts.Interfaces;
 using HeuristicLab.Hive.Contracts.BusinessObjects;
 using HeuristicLab.Hive.Contracts;
+using HeuristicLab.Security.Contracts.Interfaces;
 
 namespace HeuristicLab.Hive.Server.Core {
   public class ServerConsoleFacade: IServerConsoleFacade {
@@ -35,19 +36,22 @@ namespace HeuristicLab.Hive.Server.Core {
     private IJobManager jobManager = 
       ServiceLocator.GetJobManager();
 
-    private String loginName = null;
+    private IPermissionManager permManager = ServiceLocator.GetPermissionManager();
 
     #region IServerConsoleFacade Members
 
+    public Guid sessionID = Guid.Empty;
+
     public Response Login(string username, string password) {
       Response resp = new Response();
-
-      loginName = username;
-
-      resp.Success = true;
-      resp.StatusMessage =
-        ApplicationConstants.RESPONSE_SERVERCONSOLE_LOGIN_SUCCESS;
-
+      sessionID = permManager.Authenticate(username, password);
+      if (sessionID == Guid.Empty)
+        resp.Success = false;
+      else {
+        resp.Success = true;
+        resp.StatusMessage =
+          ApplicationConstants.RESPONSE_SERVERCONSOLE_LOGIN_SUCCESS;
+      }
       return resp;
     }
 
@@ -56,27 +60,45 @@ namespace HeuristicLab.Hive.Server.Core {
     #region IClientManager Members
 
     public ResponseList<ClientInfo> GetAllClients() {
-      return clientManager.GetAllClients();
+      if (hasPermission(PermissiveSecurityAction.List_AllClients))
+        return clientManager.GetAllClients();
+      else
+        throw new PermissionException();
     }
 
     public ResponseList<ClientGroup> GetAllClientGroups() {
-      return clientManager.GetAllClientGroups();
+      if (hasPermission(PermissiveSecurityAction.List_AllClientGroups))
+        return clientManager.GetAllClientGroups();
+      else
+        throw new PermissionException();
     }
 
     public ResponseList<UpTimeStatistics> GetAllUpTimeStatistics() {
-      return clientManager.GetAllUpTimeStatistics();
+      if (hasPermission(PermissiveSecurityAction.Show_Statistics))
+        return clientManager.GetAllUpTimeStatistics();
+      else
+        throw new PermissionException();
     }
 
     public Response AddClientGroup(ClientGroup clientGroup) {
-      return clientManager.AddClientGroup(clientGroup);
+      if (hasPermission(PermissiveSecurityAction.Add_ClientGroup))
+        return clientManager.AddClientGroup(clientGroup);
+      else
+        throw new PermissionException();
     }
 
     public Response AddResourceToGroup(Guid clientGroupId, Resource resource) {
-      return clientManager.AddResourceToGroup(clientGroupId, resource);
+      if (hasPermission(PermissiveSecurityAction.Add_Resource))
+        return clientManager.AddResourceToGroup(clientGroupId, resource);
+      else
+        throw new PermissionException();
     }
 
     public Response DeleteResourceFromGroup(Guid clientGroupId, Guid resourceId) {
-      return clientManager.DeleteResourceFromGroup(clientGroupId, resourceId);
+      if (hasPermission(PermissiveSecurityAction.Delete_Resource))
+        return clientManager.DeleteResourceFromGroup(clientGroupId, resourceId);
+      else
+        throw new PermissionException();
     }
 
     #endregion
@@ -84,32 +106,76 @@ namespace HeuristicLab.Hive.Server.Core {
     #region IJobManager Members
 
     public ResponseList<HeuristicLab.Hive.Contracts.BusinessObjects.Job> GetAllJobs() {
-      return jobManager.GetAllJobs();
+      if (hasPermission(PermissiveSecurityAction.Get_AllJobs))
+        return jobManager.GetAllJobs();
+      else
+        throw new PermissionException();
     }
+
     public ResponseObject<Job> AddNewJob(Job job) {
-      return jobManager.AddNewJob(job);
+      if (hasPermission(PermissiveSecurityAction.Add_Job))
+        return jobManager.AddNewJob(job);
+      else
+        throw new PermissionException();
     }
 
     public ResponseObject<JobResult> GetLastJobResultOf(Guid jobId, bool requested) {
-      return jobManager.GetLastJobResultOf(jobId, requested);
-    }
-
-    public Response RemoveJob(Guid jobId) {
-      return jobManager.RemoveJob(jobId);
-    }
-
-    public Response RequestSnapshot(Guid jobId) {
-      return jobManager.RequestSnapshot(jobId);
-    }
-
-    public Response AbortJob(Guid jobId) {
-      return jobManager.AbortJob(jobId);
+      if (hasPermission(PermissiveSecurityAction.Get_LastJobResult))
+        return jobManager.GetLastJobResultOf(jobId, requested);
+      else
+        throw new PermissionException();
     }
 
     public ResponseObject<List<JobResult>> GetAllJobResults(Guid jobId) {
-      throw new NotImplementedException();
+      if (hasPermission(PermissiveSecurityAction.Get_AllJobResults))
+        return jobManager.GetAllJobResults(jobId);
+      else
+        throw new PermissionException();
+    }
+
+    public Response RemoveJob(Guid jobId) {
+      if (hasPermission(PermissiveSecurityAction.Remove_Job))
+        return jobManager.RemoveJob(jobId);
+      else
+        throw new PermissionException();
+    }
+
+    public Response RequestSnapshot(Guid jobId) {
+      if (hasPermission(PermissiveSecurityAction.Request_Snapshot))
+        return jobManager.RequestSnapshot(jobId);
+      else
+        throw new PermissionException();
+    }
+
+    public Response AbortJob(Guid jobId) {
+      if (hasPermission(PermissiveSecurityAction.Abort_Job))
+        return jobManager.AbortJob(jobId);
+      else
+        throw new PermissionException();
+    }
+
+    private bool hasPermission(Guid action) {
+      return true;
+      /*
+      if (sessionID == Guid.Empty)
+        throw new Exception("sessionID is not set! Please check if user is successfully logged on!");
+      return permManager.CheckPermission(sessionID, action, Guid.Empty);
+       */
+    }
+
+    public class PermissionException : Exception {
+      public PermissionException()
+        : base("Current user has insufficent rights for this action!") {
+      }
+
+      public PermissionException(string msg)
+        : base(msg) {
+      }
+
+
     }
 
     #endregion
+
   }
 }
